# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayCat API
  description: |
    Unified Payment & Subscription Management System.
    A RevenueCat-compatible API for managing in-app purchases across iOS, Android, and Web (Stripe).
  version: 1.0.0
  contact:
    name: PayCat Support
  license:
    name: MIT
servers:
  - url: https://your-worker.workers.dev
    description: Production
  - url: http://localhost:8787
    description: Local Development
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/subscribers/{app_user_id}:
    get:
      tags: [Subscribers]
      summary: Get subscriber
      description: Get subscriber information including subscriptions and entitlements
      parameters:
        - name: app_user_id
          in: path
          required: true
          schema:
            type: string
          description: Your app's user ID
      responses:
        '200':
          description: Subscriber found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Subscribers]
      summary: Delete subscriber
      description: Delete subscriber and all associated data (GDPR)
      parameters:
        - name: app_user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subscriber deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
  /v1/subscribers/{app_user_id}/attributes:
    post:
      tags: [Subscribers]
      summary: Update subscriber attributes
      description: Set custom attributes on a subscriber
      parameters:
        - name: app_user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                attributes:
                  type: object
                  additionalProperties: true
                  example:
                    $email: "user@example.com"
                    $displayName: "John Doe"
                    custom_field: "value"
      responses:
        '200':
          description: Attributes updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for your app (pk_live_xxx or pk_test_xxx)
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin session token
  schemas:
...
    SubscriberResponse:
      type: object
      properties:
        subscriber:
          type: object
          properties:
            original_app_user_id:
              type: string
            first_seen:
              type: string
              format: date-time
            last_seen:
              type: string
              format: date-time
            subscriptions:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/Subscription'
            entitlements:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/Entitlement'
            attributes:
              type: object
    Subscription:
      type: object
      properties:
        platform:
          type: string
          enum: [ios, android, stripe]
        product_id:
          type: string
        status:
          type: string
          enum: [active, expired, cancelled, grace_period, billing_retry]
        purchase_date:
          type: string
          format: date-time
        expires_date:
          type: string
          format: date-time
        is_sandbox:
          type: boolean
        is_trial_period:
          type: boolean
        will_renew:
          type: boolean
        grace_period_expires_date:
          type: string
          format: date-time
    Entitlement:
      type: object
      properties:
        is_active:
          type: boolean
        product_identifier:
          type: string
        expires_date:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: not_found
              message: "Resource not found"